moving around some files to get everything in toolbar
