+++ src/components/notebook-header.jsx
@@ -1,19 +1,12 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 
-import { connect } from 'react-redux'
-
-import PresentationModeToolbar from './presentation-mode-toolbar'
+import PresentationModeToolbar from './toolbar/presentation-mode-toolbar'
 import EditorModeToolbar from './toolbar/editor-mode-toolbar'
 
 import DeclaredVariablesPane from './toolbar/declared-variables-pane'
 import HistoryPane from './toolbar/history-pane'
 
-export class NotebookHeader extends React.Component {
-  static propTypes = {
-    viewMode: PropTypes.oneOf(['editor', 'presentation']),
-  }
-
+export default class NotebookHeader extends React.Component {
   render() {
     return (
       <div className="notebook-header">
@@ -26,10 +19,3 @@ export class NotebookHeader extends React.Component {
     )
   }
 }
-
-function mapStateToProps(state) {
-  return {
-    viewMode: state.viewMode,
-  }
-}
-export default connect(mapStateToProps)(NotebookHeader)
